refactor(admin): retire the language surfaces the panel replaced
`fallow dead-code --trace` finds TranslationStatus with no references at
all, and it was the only thing keeping LanguageStatusPills alive outside a
test. LanguageSwitcher had already been left with nothing but its own test.
All three answered "which languages are translated, and which am I in" from
three places; the language control and the language panel answer it from
one, so they go.

Two things they held are still needed and were not theirs to own: the shape
of a `_translations` entry, and the count of how many languages are
translated. Both move to translation-meta, so retiring a surface cannot
take a shared derivation with it, and two callers cannot start counting
separately.

Removed with them: 11 tests covering only the deleted components. The
counts helper keeps its own tests, which never went through a component.
TranslationCompletenessBadge stays — the list still renders it.
